Rename plugin manager test method
