+++ core/src/main/java/io/lonmstalker/tgkit/core/bot/BotDataSourceFactory.java
@@ -6,7 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import io.lonmstalker.tgkit.core.exception.BotApiException;
-import io.lonmstalker.tgkit.core.utils.TokenCipher;
+import io.lonmstalker.tgkit.core.crypto.TokenCipher;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.telegram.telegrambots.bots.DefaultBotOptions;
